requirements.txt
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
joblib
jupyter
psutil
scapy
boto3
python-dotenv